+++ frontend/src/app/pages/exercises/exercise/shared/simulated-region-overview/simulated-region-overview.module.ts
@@ -23,7 +23,6 @@ import { SimulatedRegionOverviewBehaviorTreatPatientsPatientDetailsComponent } f
 import { WithDollarPipe } from './tabs/general-tab/utils/with-dollar';
 import { PersonnelTypeToGermanAbbreviationPipe } from './tabs/behavior-tab/utils/personnel-type-to-german-abbreviation.pipe';
 import { SimulatedRegionsModalComponent } from './simulated-regions-modal/simulated-regions-modal.component';
-import { SimulatedRegionNameComponent } from './simulated-region-name/simulated-region-name.component';
 import { SimulatedRegionOverviewPatientsTabComponent } from './tabs/patients-tab/simulated-region-overview-patients-tab/simulated-region-overview-patients-tab.component';
 import { SelectPatientService } from './select-patient.service';
 import { RadiogramListComponent } from './radiogram-list/radiogram-list.component';
@@ -64,7 +63,6 @@ import { ManagePatientTransportToHospitalSettingsEditorComponent } from './tabs/
 
 @NgModule({
     declarations: [
-        SimulatedRegionNameComponent,
         SimulatedRegionOverviewGeneralComponent,
         SimulatedRegionOverviewBehaviorTabComponent,
         SimulatedRegionOverviewGeneralTabComponent,