+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project does **not** adhere to [Semantic Versioning](https://semver.org
 
 ### Added
 
+-   Statistics can now also be restricted to simulated regions.
 -   Add functionality to create logs for statistics.
 
 ## [0.6.0] - 2023-05-17